+++ crates/messaging/affinidi-messaging-sdk/src/messages/get.rs
@@ -1,7 +1,7 @@
 use super::GetMessagesRequest;
 use crate::{
     ATM,
-    errors::{ATMError, HttpStatusError},
+    errors::{ATMError, check_response},
     messages::{GetMessagesResponse, SuccessResponse},
     profiles::ATMProfile,
 };
@@ -59,7 +59,7 @@ impl ATM {
                 })?;
 
             debug!("API response: status({})", res.status());
-            let body = HttpStatusError::check_response("get messages", res).await?;
+            let body = check_response("get messages", res).await?;
 
             let body = serde_json::from_str::<SuccessResponse<GetMessagesResponse>>(&body)
                 .map_err(|e| {