+++ crates/messaging/affinidi-messaging-sdk/src/protocols/oob_discovery.rs
@@ -6,7 +6,7 @@ Used to help discover 3rd party DID's while protecting your own privacy.
 
 use crate::{
     ATM,
-    errors::{ATMError, HttpStatusError},
+    errors::{ATMError, check_response},
     messages::{GenericDataStruct, SuccessResponse},
     profiles::ATMProfile,
 };
@@ -96,7 +96,7 @@ impl OOBDiscovery {
             })?;
 
         debug!("API response: status({})", res.status());
-        let body = HttpStatusError::check_response("create OOB invitation", res).await?;
+        let body = check_response("create OOB invitation", res).await?;
 
         let body = serde_json::from_str::<SuccessResponse<OOBInviteResponse>>(&body)
             .ok()
@@ -129,7 +129,7 @@ impl OOBDiscovery {
             })?;
 
         debug!("API response: status({})", res.status());
-        let body = HttpStatusError::check_response("retrieve OOB invitation", res).await?;
+        let body = check_response("retrieve OOB invitation", res).await?;
 
         let body = serde_json::from_str::<SuccessResponse<String>>(&body).map_err(|e| {
             ATMError::TransportError(format!("Couldn't parse OOB invitation response: {e}"))
@@ -207,7 +207,7 @@ impl OOBDiscovery {
             })?;
 
         debug!("API response: status({})", res.status());
-        let body = HttpStatusError::check_response("delete OOB invitation", res).await?;
+        let body = check_response("delete OOB invitation", res).await?;
 
         let body = serde_json::from_str::<SuccessResponse<String>>(&body)
             .ok()